# api-definitions/food-api/target/classes/api.yaml
openapi: 3.0.0
info:
  title: Food API
  version: 1.0.0
  description: Food API

servers:
  - url: http://localhost:8080/api
    description: This is food api

paths:
  /foodList:
    get:
      summary: Returns a list of all food currently available in the restaurant
      operationId: getFoodList
      responses:
        "200":
          description: All available food
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodDTO"
  /food/{id}:
    get:
      summary: Get food item by id
      operationId: getFoodById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/GivenId"
      responses:
        "200":
          description: Get food item by id
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodDTO"
    delete:
      summary: Delete food by id
      operationId: deleteFoodById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Food item deleted
  /food:
    post:
      summary: Add new food to the list
      operationId: addNewFood
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodDTO"
      responses:
        "201":
          description: Food added
  /orderList:
    get:
      summary: Returns a list of all customer order items being currently in the basket
      operationId: getOrder
      responses:
        "200":
          description: All customer order items from the basket
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerOrderDTO"
  /orderItem:
    post:
      summary: Add food to customer order basket
      operationId: addItemToOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerOrderDTO"
      responses:
        "201":
          description: Add food to customer order basket
  /orderItem/{id}:
    delete:
      summary: Delete item from customer order basket
      operationId: deleteItemFromOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/GivenId"
      responses:
        "200":
          description: Food item deleted from customer order basket
  /order/truncate:
    delete:
      summary: Truncate customer order basket
      operationId: truncateOrderTable
      responses:
        "200":
          description: Truncated customer order basket


components:
  schemas:
    GivenId:
      type: integer
      format: int64
    FoodDTO:
      type: object
      required:
        - id
        - name
        - ingredients
        - price
        - emoji
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          nullable: false
        ingredients:
          type: string
          nullable: false
        price:
          type: number
          minimum: 0
          nullable: false
        emoji:
          type: string
          nullable: false
    CustomerOrderDTO:
      type: object
      required:
        - customerOrderId
        - name
        - price
        - foodId
      properties:
        customerOrderId:
          type: integer
          format: int64
        name:
          type: string
          nullable: false
        price:
          type: number
          minimum: 0
          nullable: false
        foodId:
          type: integer
          format: int64